pyproject.toml
[project]
name = "gen-ai-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "anthropic>=0.52.0",
    "cython>=3.2.4",
    "fastapi[standard]>=0.129.0",
    "httpx>=0.28.1",
    "bcrypt>=4.0.0",
    "psycopg2-binary>=2.9.10",
    "pytest>=9.0.2",
    "python-jose[cryptography]>=3.4.0",
    "scalar-fastapi>=1.6.2",
    "sentry-sdk>=2.52.0",
    "setuptools>=82.0.0",
    "sqlalchemy>=2.0.46",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "ruff>=0.15.1",
    "ty>=0.0.17",
]

[tool.ruff.lint]
select = [
    "ANN201",  # Missing return type annotation for public function
]
